little revamp

+++ web/app/routes.py
@@ -7,11 +7,25 @@ from .scripts import draw_house
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
 predict = 0
+form = {
+    '3d_street': ['Sijslostraat'], 
+    '3d_num': ['104'],
+    'Locality': ['8020'],
+    '3d_city': ['OOSTKAMP'],
+    'Type of property': ['house'],
+    'Number of facades': ['4'],
+    'Number of rooms': ['4'],
+    'Surface of the land': ['457']
+    }
+is_loading = False
 
 @app.route('/', methods=['POST','GET'])
 @app.route('/index', methods=['POST','GET'])
 def index():
-    global predict
+    global predict, form
+    files = glob.glob(f'{dir_path}/static/3d-models/*')
+    nmesh = len(files)
+    
     if request.method == 'POST':
         try:
             dict = request.form.to_dict(flat=False)
@@ -20,24 +34,31 @@ def index():
             predict = ML.to_predict.to_predict(dict)
             # [['ertert'], ['34'], ['Bruxelles']]
             files = glob.glob(f'{dir_path}/static/3d-models/*')
-            for f in files:
-                os.remove(f)
-            draw_house.draw_houses(dict3d[0][0] + ' ' + dict3d[1][0], dict3d[2][0].upper())
+            nmesh = len(files)
+            # for f in files:
+            #     os.remove(f)
+            # draw_house.draw_houses(dict3d[0][0] + ' ' + dict3d[1][0], dict3d[2][0].upper())
             return redirect('/')
-            # return str(dict3d[2][0].upper())
         except ValueError:
             raise
     else:
         response = RequestWeather().request()
-        return render_template('index.html', title='3D Houses', weather=response, predict=predict)
+        return render_template('index.html', title='3D Houses', weather=response, predict=predict, nmesh=nmesh, form=form, is_loading=is_loading)
 
 # POST request to this endpoint(route) results in the number of votes after upvoting
 
-@app.route("/up", methods=["POST"])
-def upvote():
-    global votes
-    votes = votes + 1
-    return str(votes)
+@app.route("/load3d", methods=["POST"])
+def load3d():
+    global is_loading
+    if request.method == 'POST':
+        dict = form
+        dict3d = {key: dict[key] for key in dict.keys() & {'3d_street', '3d_num', '3d_city'}}
+        dict3d = list(dict3d.keys())
+        files = glob.glob(f'{dir_path}/static/3d-models/*')
+        for f in files:
+            os.remove(f)
+        draw_house.draw_houses(dict3d[0][0] + ' ' + dict3d[1][0], dict3d[2][0].upper())
+    return redirect('/')
 
 # POST request to this endpoint(route) results in the number of votes after downvoting
 
@@ -48,11 +69,11 @@ def downvote():
         votes = votes - 1
     return str(votes)
 
-@app.route("/predict", methods=["POST", "GET"])
-def submit():
-    if request.method == 'POST':
-        dict = request.form.to_dict(flat=False)
-        return dict
+# @app.route("/predict", methods=["POST", "GET"])
+# def submit():
+#     if request.method == 'POST':
+#         dict = request.form.to_dict(flat=False)
+#         return dict
 
 if __name__ == '__main__':
     app.run(host="0.0.0.0", port=5000, debug=True)
